Use Guice for dependency injection for all "singleton" objects
For objects that only have one instance per project, we can very easily
construct dependencies using Guice.  This means that, instead of writing
a bunch of duplicate constructors like this:

    private final EventBus eventBus;

    public MyClass(EventBus eventBus) {
        checkNotNull(eventBus, "Event Bus was null");
        this.eventBus = eventBus;
    }

...we can write:

    @Inject private EventBus eventBus;

Since a lot of classes depend on having an instance of one ore more core
classes (like Pipeline, Palette, etc...), these can also be injected,
and they automaticaly get their dependencies as well.

This is *especially* useful for JavaFX controllers.  Our current
implementation has controllers also act as roots, and they
manually load themselves from FXML and set themselves as the controller
and root in the constructor.  The only reason for this is so that we can
pass arguments to different views.

This is what a current (fictional) view in GRIP looks like:

    public class MyPipelineAndPaletteView extends HBox {
        private final EventBus eventBus;
        private final Pipeline pipeline;
        private final Palette palette;
        private final OtherGRIPView otherGripView;

        public MyPipelineAndPaletteView(EventBus eventBus, Pipeline pipeline, Palette palette) {
            this.eventBus = checkNotNull(eventBus);
            this.pipeline = checkNotNull(pipeline);
            this.palette = checkNotNull(palette);

            try {
                FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("MyPipelineAndPalette.fxml"));
                fxmlLoader.setRoot(this);
                fxmlLoade.setController(this);
                fxmlLoader.load();
            } catch (IOException e) {
                throw new RuntimeException(e);
            }

            this.otherGripView = new OtherGRIPView(eventBus, pipeline);
            this.getChildren().add(this.otherGripView);

            this.eventBus.post(new SomethingHappeningEvent(this));
        }

        @FXML
        public void onSomeButtonClick() {
            this.eventBus.post(new SomethingElseHappeningEvent(this));
        }
    }

This isn't very good MVC - some ideas from the view and controller are
expressed in the same class.  Really, the view should be entirely
defined in the FXML, but this pattern doesn't allow that.  This also
makes it hard to add additional dependencies to views - if we wanted the
pipeline view to have access to the palette, we would have to modify
wherever the pipeline view is created.

With dependency injection, the above class would look like this:

    public class MyPipelineAndPaletteController {
        @Inject private EventBus eventBus;
        @Inject private Pipeline pipeline;
        @Inject private Palette palette;

        public void initialize() {
            this.eventBus.post(new SomethingHapenningEvent(this));
        }

        @FXML
        public void onSomeButtonClick() {
            this.eventBus.post(new SomethingElseHappeningEvent(this));
        }
    }